Verbose warning message, if track cannot be loaded. (Fixes #3/2)
The problem occurs, whenever autoplay loads a track, which
backend has not initialized yet (most notable with "remote
players", such as Spotify, YTMusic and similar). Previously the
user only saw an index error in the log file. Now all the
tracks, which could not be loaded are mentioned with possible
reasons. In the previous version, Autoplay also stopped playing,
now it starts playing the next available track.